[SimpleJobQueue] Delete "hybrid" RDB capability
